@@ -1,8 +1,15 @@
+import io
+import json
 import math
 import os
-from typing import Tuple
+import shutil
+from dataclasses import asdict
 
+import einops
 import torch
+from datasets import Array2D, Dataset, Features, concatenate_datasets, load_from_disk
+from huggingface_hub import HfApi
+from jaxtyping import Float
 from tqdm import tqdm
 
 from sae_lens.config import DTYPE_MAP, CacheActivationsRunnerConfig
@@ -11,7 +18,6 @@ from sae_lens.training.activations_store import ActivationsStore
 
 
 class CacheActivationsRunner:
-
     def __init__(self, cfg: CacheActivationsRunnerConfig):
         self.cfg = cfg
         self.model = load_model(
@@ -24,8 +30,18 @@ class CacheActivationsRunner:
             self.model,
             cfg,
         )
-
-        self.file_extension = "safetensors"
+        ctx_size = _get_sliced_context_size(self.cfg)
+        self.features = Features(
+            {
+                f"{self.cfg.hook_name}": Array2D(
+                    shape=(ctx_size, self.cfg.d_in), dtype=self.cfg.dtype
+                )
+            }
+        )
+        self.tokens_in_buffer = (
+            self.cfg.n_batches_in_buffer * self.cfg.store_batch_size_prompts * ctx_size
+        )
+        self.n_buffers = math.ceil(self.cfg.training_tokens / self.tokens_in_buffer)
 
     def __str__(self):
         """
@@ -40,31 +56,48 @@ class CacheActivationsRunner:
             if isinstance(self.cfg.dtype, torch.dtype)
             else DTYPE_MAP[self.cfg.dtype].itemsize
         )
-        tokens_in_buffer = (
-            self.cfg.n_batches_in_buffer
-            * self.cfg.store_batch_size_prompts
-            * self.cfg.context_size
-        )
         total_training_tokens = self.cfg.training_tokens
         total_disk_space_gb = total_training_tokens * bytes_per_token / 10**9
 
         return (
             f"Activation Cache Runner:\n"
             f"Total training tokens: {total_training_tokens}\n"
-            f"Number of buffers: {math.ceil(total_training_tokens / tokens_in_buffer)}\n"
-            f"Tokens per buffer: {tokens_in_buffer}\n"
+            f"Number of buffers: {self.n_buffers}\n"
+            f"Tokens per buffer: {self.tokens_in_buffer}\n"
             f"Disk space required: {total_disk_space_gb:.2f} GB\n"
             f"Configuration:\n"
             f"{self.cfg}"
         )
 
     @torch.no_grad()
-    def run(self):
+    def _create_shard(
+        self,
+        buffer: Float[torch.Tensor, "(bs context_size) num_layers d_in"],
+    ) -> Dataset:
+        hook_names = [self.cfg.hook_name]  # allow multiple hooks in future
 
+        buffer = einops.rearrange(
+            buffer,
+            "(bs context_size) num_layers d_in -> num_layers bs context_size d_in",
+            bs=self.cfg.n_batches_in_buffer * self.cfg.store_batch_size_prompts,
+            context_size=_get_sliced_context_size(self.cfg),
+            d_in=self.cfg.d_in,
+            num_layers=len(hook_names),
+        )
+        shard = Dataset.from_dict(
+            {hook_name: act for hook_name, act in zip(hook_names, buffer)},
+            features=self.features,
+        )
+        return shard
+
+    @torch.no_grad()
+    def run(self) -> Dataset:
         new_cached_activations_path = self.cfg.new_cached_activations_path
+        assert new_cached_activations_path is not None
+
+        ### Paths setup
 
         # if the activations directory exists and has files in it, raise an exception
-        assert new_cached_activations_path is not None
         if os.path.exists(new_cached_activations_path):
             if len(os.listdir(new_cached_activations_path)) > 0:
                 raise Exception(
@@ -73,87 +106,83 @@ class CacheActivationsRunner:
         else:
             os.makedirs(new_cached_activations_path)
 
+        ### Create temporary sharded datasets
+
         print(f"Started caching {self.cfg.training_tokens} activations")
-        tokens_per_buffer = (
-            self.cfg.store_batch_size_prompts
-            * self.cfg.context_size
-            * self.cfg.n_batches_in_buffer
-        )
 
-        n_buffers = math.ceil(self.cfg.training_tokens / tokens_per_buffer)
+        for i in tqdm(range(self.n_buffers), desc="Caching activations"):
+            try:
+                # num activations in a single shard: n_batches_in_buffer * store_batch_size_prompts
+                buffer = self.activations_store.get_buffer(
+                    self.cfg.n_batches_in_buffer, shuffle=self.cfg.shuffle
+                )
+                shard = self._create_shard(buffer)
+                shard.save_to_disk(
+                    f"{new_cached_activations_path}/shard_{i}", num_shards=1
+                )
+                del buffer, shard
 
-        for i in tqdm(range(n_buffers), desc="Caching activations"):
-            buffer = self.activations_store.get_buffer(self.cfg.n_batches_in_buffer)
-            self.activations_store.save_buffer(
-                buffer, f"{new_cached_activations_path}/{i}.safetensors"
+            except StopIteration:
+                print(
+                    f"Warning: Ran out of samples while filling the buffer at batch {i} before reaching {self.n_buffers} batches. No more caching will occur."
+                )
+                break
+
+        ### Concat sharded datasets together, shuffle and push to hub
+
+        # mem mapped
+        dataset_shard_paths = [
+            f"{new_cached_activations_path}/shard_{i}" for i in range(self.n_buffers)
+        ]
+        dataset_shards = [
+            Dataset.load_from_disk(shard_path) for shard_path in dataset_shard_paths
+        ]
+
+        print("Concatenating shards...")
+        dataset = concatenate_datasets(dataset_shards)
+
+        if self.cfg.shuffle:
+            print("Shuffling...")
+            dataset = dataset.shuffle(seed=self.cfg.seed)
+
+        dataset.save_to_disk(new_cached_activations_path)
+
+        for shard_path in dataset_shard_paths:
+            shutil.rmtree(shard_path)
+
+        dataset = load_from_disk(new_cached_activations_path)
+        assert isinstance(dataset, Dataset)
+
+        if self.cfg.hf_repo_id:
+            print("Pushing to hub...")
+            dataset.push_to_hub(
+                repo_id=self.cfg.hf_repo_id,
+                num_shards=self.cfg.hf_num_shards or self.n_buffers,
+                private=self.cfg.hf_is_private_repo,
+                revision=self.cfg.hf_revision,
             )
 
-            del buffer
+            meta_io = io.BytesIO()
+            meta_contents = json.dumps(
+                asdict(self.cfg), indent=2, ensure_ascii=False
+            ).encode("utf-8")
+            meta_io.write(meta_contents)
+            meta_io.seek(0)
 
-            if i % self.cfg.shuffle_every_n_buffers == 0 and i > 0:
-                # Shuffle the buffers on disk
+            api = HfApi()
+            api.upload_file(
+                path_or_fileobj=meta_io,
+                path_in_repo="cache_activations_runner_cfg.json",
+                repo_id=self.cfg.hf_repo_id,
+                repo_type="dataset",
+                commit_message="Add cache_activations_runner metadata",
+            )
 
-                # Do random pairwise shuffling between the last shuffle_every_n_buffers buffers
-                for _ in range(self.cfg.n_shuffles_with_last_section):
-                    self.shuffle_activations_pairwise(
-                        new_cached_activations_path,
-                        buffer_idx_range=(i - self.cfg.shuffle_every_n_buffers, i),
-                    )
+        return dataset
 
-                # Do more random pairwise shuffling between all the buffers
-                for _ in range(self.cfg.n_shuffles_in_entire_dir):
-                    self.shuffle_activations_pairwise(
-                        new_cached_activations_path,
-                        buffer_idx_range=(0, i),
-                    )
 
-        # More final shuffling (mostly in case we didn't end on an i divisible by shuffle_every_n_buffers)
-        if n_buffers > 1:
-            for _ in tqdm(range(self.cfg.n_shuffles_final), desc="Final shuffling"):
-                self.shuffle_activations_pairwise(
-                    new_cached_activations_path,
-                    buffer_idx_range=(0, n_buffers),
-                )
-
-    @torch.no_grad()
-    def shuffle_activations_pairwise(
-        self, datapath: str, buffer_idx_range: Tuple[int, int]
-    ):
-        """
-        Shuffles two buffers on disk.
-        """
-        assert (
-            buffer_idx_range[0] < buffer_idx_range[1] - 1
-        ), "buffer_idx_range[0] must be smaller than buffer_idx_range[1] by at least 1"
-
-        buffer_idx1 = torch.randint(
-            buffer_idx_range[0], buffer_idx_range[1], (1,)
-        ).item()
-        buffer_idx2 = torch.randint(
-            buffer_idx_range[0], buffer_idx_range[1], (1,)
-        ).item()
-        while buffer_idx1 == buffer_idx2:  # Make sure they're not the same
-            buffer_idx2 = torch.randint(
-                buffer_idx_range[0], buffer_idx_range[1], (1,)
-            ).item()
-
-        buffer1 = self.activations_store.load_buffer(
-            f"{datapath}/{buffer_idx1}.{self.file_extension}"
-        )
-        buffer2 = self.activations_store.load_buffer(
-            f"{datapath}/{buffer_idx2}.{self.file_extension}"
-        )
-        joint_buffer = torch.cat([buffer1, buffer2])
-
-        # Shuffle them
-        joint_buffer = joint_buffer[torch.randperm(joint_buffer.shape[0])]
-        shuffled_buffer1 = joint_buffer[: buffer1.shape[0]]
-        shuffled_buffer2 = joint_buffer[buffer1.shape[0] :]
-
-        # Save them back
-        self.activations_store.save_buffer(
-            shuffled_buffer1, f"{datapath}/{buffer_idx1}.{self.file_extension}"
-        )
-        self.activations_store.save_buffer(
-            shuffled_buffer2, f"{datapath}/{buffer_idx2}.{self.file_extension}"
-        )
+def _get_sliced_context_size(cfg: CacheActivationsRunnerConfig) -> int:
+    context_size = cfg.context_size
+    if cfg.seqpos_slice:
+        context_size = len(range(context_size)[slice(*cfg.seqpos_slice)])
+    return context_size
