@@ -1,4 +1,5 @@
 import json
+import logging
 import os
 import signal
 from typing import Any, cast
@@ -8,7 +9,7 @@ import wandb
 from safetensors.torch import save_file
 from transformer_lens.hook_points import HookedRootModule
 
-from sae_lens.config import LanguageModelSAERunnerConfig
+from sae_lens.config import HfDataset, LanguageModelSAERunnerConfig
 from sae_lens.load_model import load_model
 from sae_lens.sae import SAE_CFG_PATH, SAE_WEIGHTS_PATH, SPARSITY_PATH
 from sae_lens.training.activations_store import ActivationsStore
@@ -35,19 +36,37 @@ class SAETrainingRunner:
     sae: TrainingSAE
     activations_store: ActivationsStore
 
-    def __init__(self, cfg: LanguageModelSAERunnerConfig):
+    def __init__(
+        self,
+        cfg: LanguageModelSAERunnerConfig,
+        override_dataset: HfDataset | None = None,
+        override_model: HookedRootModule | None = None,
+    ):
+        if override_dataset is not None:
+            logging.warning(
+                f"You just passed in a dataset which will override the one specified in your configuration: {cfg.dataset_path}. As a consequence this run will not be reproducible via configuration alone."
+            )
+        if override_model is not None:
+            logging.warning(
+                f"You just passed in a model which will override the one specified in your configuration: {cfg.model_name}. As a consequence this run will not be reproducible via configuration alone."
+            )
+
         self.cfg = cfg
 
-        self.model = load_model(
-            self.cfg.model_class_name,
-            self.cfg.model_name,
-            device=self.cfg.device,
-            model_from_pretrained_kwargs=self.cfg.model_from_pretrained_kwargs,
-        )
+        if override_model is None:
+            self.model = load_model(
+                self.cfg.model_class_name,
+                self.cfg.model_name,
+                device=self.cfg.device,
+                model_from_pretrained_kwargs=self.cfg.model_from_pretrained_kwargs,
+            )
+        else:
+            self.model = override_model
 
         self.activations_store = ActivationsStore.from_config(
             self.model,
             self.cfg,
+            override_dataset=override_dataset,
         )
 
         if self.cfg.from_pretrained_path is not None:
@@ -70,6 +89,7 @@ class SAETrainingRunner:
         if self.cfg.log_to_wandb:
             wandb.init(
                 project=self.cfg.wandb_project,
+                entity=self.cfg.wandb_entity,
                 config=cast(Any, self.cfg),
                 name=self.cfg.run_name,
                 id=self.cfg.wandb_id,
@@ -87,7 +107,8 @@ class SAETrainingRunner:
         sae = self.run_trainer_with_interruption_handling(trainer)
 
         if self.cfg.log_to_wandb:
-            wandb.finish()
+            # remove this type ignore comment after https://github.com/wandb/wandb/issues/8248 is resolved
+            wandb.finish()  # type: ignore
 
         return sae
 
@@ -188,8 +209,12 @@ class SAETrainingRunner:
         save_file(log_feature_sparsities, log_feature_sparsity_path)
 
         if trainer.cfg.log_to_wandb and os.path.exists(log_feature_sparsity_path):
+            # Avoid wandb saving errors such as:
+            #   ValueError: Artifact name may only contain alphanumeric characters, dashes, underscores, and dots. Invalid name: sae_google/gemma-2b_etc
+            sae_name = self.sae.get_name().replace("/", "__")
+
             model_artifact = wandb.Artifact(
-                f"{self.sae.get_name()}",
+                sae_name,
                 type="model",
                 metadata=dict(trainer.cfg.__dict__),
             )
@@ -197,14 +222,16 @@ class SAETrainingRunner:
             model_artifact.add_file(f"{path}/{SAE_WEIGHTS_PATH}")
             model_artifact.add_file(f"{path}/{SAE_CFG_PATH}")
 
-            wandb.log_artifact(model_artifact, aliases=wandb_aliases)
+            # remove this type ignore comment after https://github.com/wandb/wandb/issues/8248 is resolved
+            wandb.log_artifact(model_artifact, aliases=wandb_aliases)  # type: ignore
 
             sparsity_artifact = wandb.Artifact(
-                f"{self.sae.get_name()}_log_feature_sparsity",
+                f"{sae_name}_log_feature_sparsity",
                 type="log_feature_sparsity",
                 metadata=dict(trainer.cfg.__dict__),
             )
             sparsity_artifact.add_file(log_feature_sparsity_path)
-            wandb.log_artifact(sparsity_artifact)
+            # remove this type ignore comment after https://github.com/wandb/wandb/issues/8248 is resolved
+            wandb.log_artifact(sparsity_artifact)  # type: ignore
 
         return checkpoint_path
