@@ -1,32 +1,33 @@
 from __future__ import annotations
 
 import contextlib
+import json
 import os
-from typing import Any, Iterator, Literal, cast
+import warnings
+from typing import Any, Generator, Iterator, Literal, cast
 
+import datasets
 import numpy as np
 import torch
-from datasets import (
-    Dataset,
-    DatasetDict,
-    IterableDataset,
-    IterableDatasetDict,
-    load_dataset,
-)
-from safetensors import safe_open
+from datasets import Dataset, DatasetDict, IterableDataset, load_dataset
+from huggingface_hub import hf_hub_download
+from huggingface_hub.utils import HfHubHTTPError
+from jaxtyping import Float
+from requests import HTTPError
 from safetensors.torch import save_file
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 from transformer_lens.hook_points import HookedRootModule
+from transformers import AutoTokenizer, PreTrainedTokenizerBase
 
 from sae_lens.config import (
     DTYPE_MAP,
     CacheActivationsRunnerConfig,
+    HfDataset,
     LanguageModelSAERunnerConfig,
 )
 from sae_lens.sae import SAE
-
-HfDataset = DatasetDict | Dataset | IterableDatasetDict | IterableDataset
+from sae_lens.tokenization_and_batching import concat_and_batch_sequences
 
 
 # TODO: Make an activation store config class to be consistent with the rest of the code.
@@ -39,7 +40,8 @@ class ActivationsStore:
     model: HookedRootModule
     dataset: HfDataset
     cached_activations_path: str | None
-    tokens_column: Literal["tokens", "input_ids", "text"]
+    cached_activation_dataset: Dataset | None = None
+    tokens_column: Literal["tokens", "input_ids", "text", "problem"]
     hook_name: str
     hook_layer: int
     hook_head_index: int | None
@@ -52,7 +54,7 @@ class ActivationsStore:
         cls,
         model: HookedRootModule,
         cfg: LanguageModelSAERunnerConfig | CacheActivationsRunnerConfig,
-        dataset: HfDataset | None = None,
+        override_dataset: HfDataset | None = None,
     ) -> "ActivationsStore":
         cached_activations_path = cfg.cached_activations_path
         # set cached_activations_path to None if we're not using cached activations
@@ -61,9 +63,15 @@ class ActivationsStore:
             and not cfg.use_cached_activations
         ):
             cached_activations_path = None
+
+        if override_dataset is None and cfg.dataset_path == "":
+            raise ValueError(
+                "You must either pass in a dataset or specify a dataset_path in your configutation."
+            )
+
         return cls(
             model=model,
-            dataset=dataset or cfg.dataset_path,
+            dataset=override_dataset or cfg.dataset_path,
             streaming=cfg.streaming,
             hook_name=cfg.hook_name,
             hook_layer=cfg.hook_layer,
@@ -82,6 +90,7 @@ class ActivationsStore:
             model_kwargs=cfg.model_kwargs,
             autocast_lm=cfg.autocast_lm,
             dataset_trust_remote_code=cfg.dataset_trust_remote_code,
+            seqpos_slice=cfg.seqpos_slice,
         )
 
     @classmethod
@@ -89,6 +98,8 @@ class ActivationsStore:
         cls,
         model: HookedRootModule,
         sae: SAE,
+        context_size: int | None = None,
+        dataset: HfDataset | str | None = None,
         streaming: bool = True,
         store_batch_size_prompts: int = 8,
         n_batches_in_buffer: int = 8,
@@ -96,15 +107,14 @@ class ActivationsStore:
         total_tokens: int = 10**9,
         device: str = "cpu",
     ) -> "ActivationsStore":
-
         return cls(
             model=model,
-            dataset=sae.cfg.dataset_path,
+            dataset=sae.cfg.dataset_path if dataset is None else dataset,
             d_in=sae.cfg.d_in,
             hook_name=sae.cfg.hook_name,
             hook_layer=sae.cfg.hook_layer,
             hook_head_index=sae.cfg.hook_head_index,
-            context_size=sae.cfg.context_size,
+            context_size=sae.cfg.context_size if context_size is None else context_size,
             prepend_bos=sae.cfg.prepend_bos,
             streaming=streaming,
             store_batch_size_prompts=store_batch_size_prompts,
@@ -115,6 +125,7 @@ class ActivationsStore:
             dataset_trust_remote_code=sae.cfg.dataset_trust_remote_code,
             dtype=sae.cfg.dtype,
             device=torch.device(device),
+            seqpos_slice=sae.cfg.seqpos_slice,
         )
 
     def __init__(
@@ -139,6 +150,7 @@ class ActivationsStore:
         model_kwargs: dict[str, Any] | None = None,
         autocast_lm: bool = False,
         dataset_trust_remote_code: bool | None = None,
+        seqpos_slice: tuple[int | None, ...] = (None,),
     ):
         self.model = model
         if model_kwargs is None:
@@ -154,12 +166,23 @@ class ActivationsStore:
             if isinstance(dataset, str)
             else dataset
         )
+
+        if isinstance(dataset, (Dataset, DatasetDict)):
+            self.dataset = cast(Dataset | DatasetDict, self.dataset)
+            n_samples = len(self.dataset)
+
+            if n_samples < total_training_tokens:
+                warnings.warn(
+                    f"The training dataset contains fewer samples ({n_samples}) than the number of samples required by your training configuration ({total_training_tokens}). This will result in multiple training epochs and some samples being used more than once."
+                )
+
         self.hook_name = hook_name
         self.hook_layer = hook_layer
         self.hook_head_index = hook_head_index
         self.context_size = context_size
         self.d_in = d_in
         self.n_batches_in_buffer = n_batches_in_buffer
+        self.half_buffer_size = n_batches_in_buffer // 2
         self.total_training_tokens = total_training_tokens
         self.store_batch_size_prompts = store_batch_size_prompts
         self.train_batch_size_tokens = train_batch_size_tokens
@@ -169,14 +192,14 @@ class ActivationsStore:
         self.dtype = DTYPE_MAP[dtype]
         self.cached_activations_path = cached_activations_path
         self.autocast_lm = autocast_lm
+        self.seqpos_slice = seqpos_slice
 
         self.n_dataset_processed = 0
-        self.iterable_dataset = iter(self.dataset)
 
         self.estimated_norm_scaling_factor = 1.0
 
         # Check if dataset is tokenized
-        dataset_sample = next(self.iterable_dataset)
+        dataset_sample = next(iter(self.dataset))
 
         # check if it's tokenized
         if "tokens" in dataset_sample.keys():
@@ -188,42 +211,151 @@ class ActivationsStore:
         elif "text" in dataset_sample.keys():
             self.is_dataset_tokenized = False
             self.tokens_column = "text"
+        elif "problem" in dataset_sample.keys():
+            self.is_dataset_tokenized = False
+            self.tokens_column = "problem"
         else:
             raise ValueError(
-                "Dataset must have a 'tokens', 'input_ids', or 'text' column."
+                "Dataset must have a 'tokens', 'input_ids', 'text', or 'problem' column."
             )
-        self.iterable_dataset = iter(self.dataset)  # Reset iterator after checking
+        if self.is_dataset_tokenized:
+            ds_context_size = len(dataset_sample[self.tokens_column])
+            if ds_context_size < self.context_size:
+                raise ValueError(
+                    f"""pretokenized dataset has context_size {ds_context_size}, but the provided context_size is {self.context_size}.
+                    The context_size {ds_context_size} is expected to be larger than or equal to the provided context size {self.context_size}."""
+                )
+            if self.context_size != ds_context_size:
+                warnings.warn(
+                    f"""pretokenized dataset has context_size {ds_context_size}, but the provided context_size is {self.context_size}. Some data will be discarded in this case.""",
+                    RuntimeWarning,
+                )
+            # TODO: investigate if this can work for iterable datasets, or if this is even worthwhile as a perf improvement
+            if hasattr(self.dataset, "set_format"):
+                self.dataset.set_format(type="torch", columns=[self.tokens_column])  # type: ignore
 
-        self.check_cached_activations_against_config()
+            if (
+                isinstance(dataset, str)
+                and hasattr(model, "tokenizer")
+                and model.tokenizer is not None
+            ):
+                validate_pretokenized_dataset_tokenizer(
+                    dataset_path=dataset, model_tokenizer=model.tokenizer
+                )
+        else:
+            warnings.warn(
+                "Dataset is not tokenized. Pre-tokenizing will improve performance and allows for more control over special tokens. See https://jbloomaus.github.io/SAELens/training_saes/#pretokenizing-datasets for more info."
+            )
+
+        self.iterable_sequences = self._iterate_tokenized_sequences()
+
+        self.cached_activation_dataset = self.load_cached_activation_dataset()
 
         # TODO add support for "mixed loading" (ie use cache until you run out, then switch over to streaming from HF)
 
-    def check_cached_activations_against_config(self):
+    def _iterate_raw_dataset(
+        self,
+    ) -> Generator[torch.Tensor | list[int] | str, None, None]:
+        """
+        Helper to iterate over the dataset while incrementing n_dataset_processed
+        """
+        for row in self.dataset:
+            # typing datasets is difficult
+            yield row[self.tokens_column]  # type: ignore
+            self.n_dataset_processed += 1
 
-        if self.cached_activations_path is not None:  # EDIT: load from multi-layer acts
-            assert self.cached_activations_path is not None  # keep pyright happy
-            # Sanity check: does the cache directory exist?
-            assert os.path.exists(
-                self.cached_activations_path
-            ), f"Cache directory {self.cached_activations_path} does not exist. Consider double-checking your dataset, model, and hook names."
+    def _iterate_raw_dataset_tokens(self) -> Generator[torch.Tensor, None, None]:
+        """
+        Helper to create an iterator which tokenizes raw text from the dataset on the fly
+        """
+        for row in self._iterate_raw_dataset():
+            tokens = (
+                self.model.to_tokens(
+                    row,
+                    truncate=False,
+                    move_to_device=True,
+                    prepend_bos=False,
+                )
+                .squeeze(0)
+                .to(self.device)
+            )
+            assert (
+                len(tokens.shape) == 1
+            ), f"tokens.shape should be 1D but was {tokens.shape}"
+            yield tokens
 
-            self.next_cache_idx = 0  # which file to open next
-            self.next_idx_within_buffer = 0  # where to start reading from in that file
-
-            # Check that we have enough data on disk
-            first_buffer = self.load_buffer(
-                f"{self.cached_activations_path}/{self.next_cache_idx}.safetensors"
+    def _iterate_tokenized_sequences(self) -> Generator[torch.Tensor, None, None]:
+        """
+        Generator which iterates over full sequence of context_size tokens
+        """
+        # If the datset is pretokenized, we will slice the dataset to the length of the context window if needed. Otherwise, no further processing is needed.
+        # We assume that all necessary BOS/EOS/SEP tokens have been added during pretokenization.
+        if self.is_dataset_tokenized:
+            for row in self._iterate_raw_dataset():
+                yield torch.tensor(
+                    row[
+                        : self.context_size
+                    ],  # If self.context_size = None, this line simply returns the whole row
+                    dtype=torch.long,
+                    device=self.device,
+                    requires_grad=False,
+                )
+        # If the dataset isn't tokenized, we'll tokenize, concat, and batch on the fly
+        else:
+            tokenizer = getattr(self.model, "tokenizer", None)
+            bos_token_id = None if tokenizer is None else tokenizer.bos_token_id
+            yield from concat_and_batch_sequences(
+                tokens_iterator=self._iterate_raw_dataset_tokens(),
+                context_size=self.context_size,
+                begin_batch_token_id=(bos_token_id if self.prepend_bos else None),
+                begin_sequence_token_id=None,
+                sequence_separator_token_id=(
+                    bos_token_id if self.prepend_bos else None
+                ),
             )
 
-            buffer_size_on_disk = first_buffer.shape[0]
-            n_buffers_on_disk = len(os.listdir(self.cached_activations_path))
+    def load_cached_activation_dataset(self) -> Dataset | None:
+        """
+        Load the cached activation dataset from disk.
 
-            # Note: we're assuming all files have the same number of tokens
-            # (which seems reasonable imo since that's what our script does)
-            n_activations_on_disk = buffer_size_on_disk * n_buffers_on_disk
-            assert (
-                n_activations_on_disk >= self.total_training_tokens
-            ), f"Only {n_activations_on_disk/1e6:.1f}M activations on disk, but total_training_tokens is {self.total_training_tokens/1e6:.1f}M."
+        - If cached_activations_path is set, returns Huggingface Dataset else None
+        - Checks that the loaded dataset has current has activations for hooks in config and that shapes match.
+        """
+        if self.cached_activations_path is None:
+            return None
+
+        assert self.cached_activations_path is not None  # keep pyright happy
+        # Sanity check: does the cache directory exist?
+        assert os.path.exists(
+            self.cached_activations_path
+        ), f"Cache directory {self.cached_activations_path} does not exist. Consider double-checking your dataset, model, and hook names."
+
+        # ---
+        # Actual code
+        activations_dataset = datasets.load_from_disk(self.cached_activations_path)
+        activations_dataset.set_format(
+            type="torch", columns=[self.hook_name], device=self.device, dtype=self.dtype
+        )
+        self.current_row_idx = 0  # idx to load next batch from
+        # ---
+
+        assert isinstance(activations_dataset, Dataset)
+
+        # multiple in hooks future
+        if not set([self.hook_name]).issubset(activations_dataset.column_names):
+            raise ValueError(
+                f"loaded dataset does not include hook activations, got {activations_dataset.column_names}"
+            )
+
+        if activations_dataset.features[self.hook_name].shape != (
+            self.context_size,
+            self.d_in,
+        ):
+            raise ValueError(
+                f"Given dataset of shape {activations_dataset.features[self.hook_name].shape} does not match context_size ({self.context_size}) and d_in ({self.d_in})"
+            )
+
+        return activations_dataset
 
     def apply_norm_scaling_factor(self, activations: torch.Tensor) -> torch.Tensor:
         return activations * self.estimated_norm_scaling_factor
@@ -236,7 +368,6 @@ class ActivationsStore:
 
     @torch.no_grad()
     def estimate_norm_scaling_factor(self, n_batches_for_norm_estimate: int = int(1e3)):
-
         norms_per_batch = []
         for _ in tqdm(
             range(n_batches_for_norm_estimate), desc="Estimating norm scaling factor"
@@ -248,10 +379,30 @@ class ActivationsStore:
 
         return scaling_factor
 
+    def shuffle_input_dataset(self, seed: int, buffer_size: int = 1):
+        """
+        This applies a shuffle to the huggingface dataset that is the input to the activations store. This
+        also shuffles the shards of the dataset, which is especially useful for evaluating on different
+        sections of very large streaming datasets. Buffer size is only relevant for streaming datasets.
+        The default buffer_size of 1 means that only the shard will be shuffled; larger buffer sizes will
+        additionally shuffle individual elements within the shard.
+        """
+        if type(self.dataset) == IterableDataset:
+            self.dataset = self.dataset.shuffle(seed=seed, buffer_size=buffer_size)
+        else:
+            self.dataset = self.dataset.shuffle(seed=seed)
+        self.iterable_dataset = iter(self.dataset)
+
+    def reset_input_dataset(self):
+        """
+        Resets the input dataset iterator to the beginning.
+        """
+        self.iterable_dataset = iter(self.dataset)
+
     @property
     def storage_buffer(self) -> torch.Tensor:
         if self._storage_buffer is None:
-            self._storage_buffer = self.get_buffer(self.n_batches_in_buffer // 2)
+            self._storage_buffer = self.get_buffer(self.half_buffer_size)
 
         return self._storage_buffer
 
@@ -261,76 +412,31 @@ class ActivationsStore:
             self._dataloader = self.get_data_loader()
         return self._dataloader
 
-    def get_batch_tokens(self, batch_size: int | None = None):
+    def get_batch_tokens(
+        self, batch_size: int | None = None, raise_at_epoch_end: bool = False
+    ):
         """
         Streams a batch of tokens from a dataset.
+
+        If raise_at_epoch_end is true we will reset the dataset at the end of each epoch and raise a StopIteration. Otherwise we will reset silently.
         """
         if not batch_size:
             batch_size = self.store_batch_size_prompts
-        context_size = self.context_size
-        device = self.device
-
-        batch_tokens = torch.zeros(
-            size=(0, context_size), device=device, dtype=torch.long, requires_grad=False
-        )
-
-        current_batch = []
-        current_length = 0
-
-        while batch_tokens.shape[0] < batch_size:
-            tokens = self._get_next_dataset_tokens()
-            token_len = tokens.shape[0]
-
-            # TODO: Fix this so that we are limiting how many tokens we get from the same context.
-            assert self.model.tokenizer is not None  # keep pyright happy
-            while token_len > 0 and batch_tokens.shape[0] < batch_size:
-                # Space left in the current batch
-                space_left = context_size - current_length
-
-                # If the current tokens fit entirely into the remaining space
-                if token_len <= space_left:
-                    current_batch.append(tokens[:token_len])
-                    current_length += token_len
-                    break
-
-                else:
-                    # Take as much as will fit
-                    current_batch.append(tokens[:space_left])
-
-                    # Remove used part, add BOS
-                    tokens = tokens[space_left:]
-                    token_len -= space_left
-
-                    # only add BOS if it's not already the first token
-                    if self.prepend_bos:
-                        bos_token_id_tensor = torch.tensor(
-                            [self.model.tokenizer.bos_token_id],
-                            device=tokens.device,
-                            dtype=torch.long,
-                        )
-                        if tokens[0] != bos_token_id_tensor:
-                            tokens = torch.cat(
-                                (
-                                    bos_token_id_tensor,
-                                    tokens,
-                                ),
-                                dim=0,
-                            )
-                            token_len += 1
-                    current_length = context_size
-
-                # If a batch is full, concatenate and move to next batch
-                if current_length == context_size:
-                    full_batch = torch.cat(current_batch, dim=0)
-                    batch_tokens = torch.cat(
-                        (batch_tokens, full_batch.unsqueeze(0)), dim=0
+        sequences = []
+        # the sequences iterator yields fully formed tokens of size context_size, so we just need to cat these into a batch
+        for _ in range(batch_size):
+            try:
+                sequences.append(next(self.iterable_sequences))
+            except StopIteration:
+                self.iterable_sequences = self._iterate_tokenized_sequences()
+                if raise_at_epoch_end:
+                    raise StopIteration(
+                        f"Ran out of tokens in dataset after {self.n_dataset_processed} samples, beginning the next epoch."
                     )
-                    current_batch = []
-                    current_length = 0
+                else:
+                    sequences.append(next(self.iterable_sequences))
 
-            # pbar.n = batch_tokens.shape[0]
-            # pbar.refresh()
-        return batch_tokens[:batch_size].to(self.model.W_E.device)
+        return torch.stack(sequences, dim=0).to(self.model.W_E.device)
 
     @torch.no_grad()
     def get_activations(self, batch_tokens: torch.Tensor):
@@ -351,105 +457,125 @@ class ActivationsStore:
             autocast_if_enabled = contextlib.nullcontext()
 
         with autocast_if_enabled:
-            layerwise_activations = self.model.run_with_cache(
+            layerwise_activations_cache = self.model.run_with_cache(
                 batch_tokens,
                 names_filter=[self.hook_name],
                 stop_at_layer=self.hook_layer + 1,
-                prepend_bos=self.prepend_bos,
+                prepend_bos=False,
                 **self.model_kwargs,
             )[1]
 
-        n_batches, n_context = batch_tokens.shape
+        layerwise_activations = layerwise_activations_cache[self.hook_name][
+            :, slice(*self.seqpos_slice)
+        ]
+
+        n_batches, n_context = layerwise_activations.shape[:2]
 
         stacked_activations = torch.zeros((n_batches, n_context, 1, self.d_in))
 
         if self.hook_head_index is not None:
-            stacked_activations[:, :, 0] = layerwise_activations[self.hook_name][
+            stacked_activations[:, :, 0] = layerwise_activations[
                 :, :, self.hook_head_index
             ]
-        elif (
-            layerwise_activations[self.hook_name].ndim > 3
-        ):  # if we have a head dimension
-            stacked_activations[:, :, 0] = layerwise_activations[self.hook_name].view(
-                n_batches, n_context, -1
-            )
+        elif layerwise_activations.ndim > 3:  # if we have a head dimension
+            try:
+                stacked_activations[:, :, 0] = layerwise_activations.view(
+                    n_batches, n_context, -1
+                )
+            except RuntimeError as e:
+                print(f"Error during view operation: {e}")
+                print("Attempting to use reshape instead...")
+                stacked_activations[:, :, 0] = layerwise_activations.reshape(
+                    n_batches, n_context, -1
+                )
         else:
-            stacked_activations[:, :, 0] = layerwise_activations[self.hook_name]
+            stacked_activations[:, :, 0] = layerwise_activations
 
         return stacked_activations
 
+    def _load_buffer_from_cached(
+        self,
+        total_size: int,
+        context_size: int,
+        num_layers: int,
+        d_in: int,
+        raise_on_epoch_end: bool,
+    ) -> Float[torch.Tensor, "(total_size context_size) num_layers d_in"]:
+        """
+        Loads `total_size` activations from `cached_activation_dataset`
+
+        The dataset has columns for each hook_name,
+        each containing activations of shape (context_size, d_in).
+
+        raises StopIteration
+        """
+        assert self.cached_activation_dataset is not None
+        # In future, could be a list of multiple hook names
+        hook_names = [self.hook_name]
+        assert set(hook_names).issubset(self.cached_activation_dataset.column_names)
+
+        if self.current_row_idx > len(self.cached_activation_dataset) - total_size:
+            self.current_row_idx = 0
+            if raise_on_epoch_end:
+                raise StopIteration
+
+        new_buffer = []
+        for hook_name in hook_names:
+            # Load activations for each hook.
+            # Usually faster to first slice dataset then pick column
+            _hook_buffer = self.cached_activation_dataset[
+                self.current_row_idx : self.current_row_idx + total_size
+            ][hook_name]
+            assert _hook_buffer.shape == (total_size, context_size, d_in)
+            new_buffer.append(_hook_buffer)
+
+        # Stack across num_layers dimension
+        # list of num_layers; shape: (total_size, context_size, d_in) -> (total_size, context_size, num_layers, d_in)
+        new_buffer = torch.stack(new_buffer, dim=2)
+        assert new_buffer.shape == (total_size, context_size, num_layers, d_in)
+
+        self.current_row_idx += total_size
+        return new_buffer.reshape(total_size * context_size, num_layers, d_in)
+
     @torch.no_grad()
-    def get_buffer(self, n_batches_in_buffer: int) -> torch.Tensor:
+    def get_buffer(
+        self,
+        n_batches_in_buffer: int,
+        raise_on_epoch_end: bool = False,
+        shuffle: bool = True,
+    ) -> torch.Tensor:
+        """
+        Loads the next n_batches_in_buffer batches of activations into a tensor and returns half of it.
+
+        The primary purpose here is maintaining a shuffling buffer.
+
+        If raise_on_epoch_end is True, when the dataset it exhausted it will automatically refill the dataset and then raise a StopIteration so that the caller has a chance to react.
+        """
         context_size = self.context_size
+        training_context_size = len(range(context_size)[slice(*self.seqpos_slice)])
         batch_size = self.store_batch_size_prompts
         d_in = self.d_in
         total_size = batch_size * n_batches_in_buffer
         num_layers = 1
 
-        if self.cached_activations_path is not None:
-            # Load the activations from disk
-            buffer_size = total_size * context_size
-            # Initialize an empty tensor with an additional dimension for layers
-            new_buffer = torch.zeros(
-                (buffer_size, num_layers, d_in),
-                dtype=self.dtype,  # type: ignore
-                device=self.device,
+        if self.cached_activation_dataset is not None:
+            return self._load_buffer_from_cached(
+                total_size, context_size, num_layers, d_in, raise_on_epoch_end
             )
-            n_tokens_filled = 0
-
-            # Assume activations for different layers are stored separately and need to be combined
-            while n_tokens_filled < buffer_size:
-                if not os.path.exists(
-                    f"{self.cached_activations_path}/{self.next_cache_idx}.safetensors"
-                ):
-                    print(
-                        "\n\nWarning: Ran out of cached activation files earlier than expected."
-                    )
-                    print(
-                        f"Expected to have {buffer_size} activations, but only found {n_tokens_filled}."
-                    )
-                    if buffer_size % self.total_training_tokens != 0:
-                        print(
-                            "This might just be a rounding error — your batch_size * n_batches_in_buffer * context_size is not divisible by your total_training_tokens"
-                        )
-                    print(f"Returning a buffer of size {n_tokens_filled} instead.")
-                    print("\n\n")
-                    new_buffer = new_buffer[:n_tokens_filled, ...]
-                    return new_buffer
-
-                activations = self.load_buffer(
-                    f"{self.cached_activations_path}/{self.next_cache_idx}.safetensors"
-                )
-                taking_subset_of_file = False
-                if n_tokens_filled + activations.shape[0] > buffer_size:
-                    activations = activations[: buffer_size - n_tokens_filled, ...]
-                    taking_subset_of_file = True
-
-                new_buffer[
-                    n_tokens_filled : n_tokens_filled + activations.shape[0], ...
-                ] = activations
-
-                if taking_subset_of_file:
-                    self.next_idx_within_buffer = activations.shape[0]
-                else:
-                    self.next_cache_idx += 1
-                    self.next_idx_within_buffer = 0
-
-                n_tokens_filled += activations.shape[0]
-
-            return new_buffer
 
         refill_iterator = range(0, batch_size * n_batches_in_buffer, batch_size)
         # Initialize empty tensor buffer of the maximum required size with an additional dimension for layers
         new_buffer = torch.zeros(
-            (total_size, context_size, num_layers, d_in),
+            (total_size, training_context_size, num_layers, d_in),
             dtype=self.dtype,  # type: ignore
             device=self.device,
         )
 
         for refill_batch_idx_start in refill_iterator:
             # move batch toks to gpu for model
-            refill_batch_tokens = self.get_batch_tokens().to(self.model.cfg.device)
+            refill_batch_tokens = self.get_batch_tokens(
+                raise_at_epoch_end=raise_on_epoch_end
+            ).to(self.model.cfg.device)
             refill_activations = self.get_activations(refill_batch_tokens)
             # move acts back to cpu
             refill_activations.to(self.device)
@@ -460,26 +586,14 @@ class ActivationsStore:
             # pbar.update(1)
 
         new_buffer = new_buffer.reshape(-1, num_layers, d_in)
-        new_buffer = new_buffer[torch.randperm(new_buffer.shape[0])]
+        if shuffle:
+            new_buffer = new_buffer[torch.randperm(new_buffer.shape[0])]
 
         # every buffer should be normalized:
         if self.normalize_activations == "expected_average_only_in":
             new_buffer = self.apply_norm_scaling_factor(new_buffer)
 
         return new_buffer
-
-    def save_buffer(self, buffer: torch.Tensor, path: str):
-        """
-        Used by cached activations runner to save a buffer to disk.
-        For reuse by later workflows.
-        """
-        save_file({"activations": buffer}, path)
-
-    def load_buffer(self, path: str) -> torch.Tensor:
-
-        with safe_open(path, framework="pt", device=str(self.device)) as f:  # type: ignore
-            buffer = f.get_tensor("activations")
-        return buffer
 
     def get_data_loader(
         self,
@@ -494,9 +608,27 @@ class ActivationsStore:
 
         batch_size = self.train_batch_size_tokens
 
+        try:
+            new_samples = self.get_buffer(
+                self.half_buffer_size, raise_on_epoch_end=True
+            )
+        except StopIteration:
+            warnings.warn(
+                "All samples in the training dataset have been exhausted, we are now beginning a new epoch with the same samples."
+            )
+            self._storage_buffer = (
+                None  # dump the current buffer so samples do not leak between epochs
+            )
+            try:
+                new_samples = self.get_buffer(self.half_buffer_size)
+            except StopIteration:
+                raise ValueError(
+                    "We were unable to fill up the buffer directly after starting a new epoch. This could indicate that there are less samples in the dataset than are required to fill up the buffer. Consider reducing batch_size or n_batches_in_buffer. "
+                )
+
         # 1. # create new buffer by mixing stored and new buffer
         mixing_buffer = torch.cat(
-            [self.get_buffer(self.n_batches_in_buffer // 2), self.storage_buffer],
+            [new_samples, self.storage_buffer],
             dim=0,
         )
 
@@ -541,34 +673,30 @@ class ActivationsStore:
     def save(self, file_path: str):
         save_file(self.state_dict(), file_path)
 
-    def _get_next_dataset_tokens(self) -> torch.Tensor:
-        device = self.device
-        if not self.is_dataset_tokenized:
-            s = next(self.iterable_dataset)[self.tokens_column]
-            tokens = (
-                self.model.to_tokens(
-                    s,
-                    truncate=False,
-                    move_to_device=True,
-                    prepend_bos=self.prepend_bos,
-                )
-                .squeeze(0)
-                .to(device)
-            )
-            assert (
-                len(tokens.shape) == 1
-            ), f"tokens.shape should be 1D but was {tokens.shape}"
-        else:
-            tokens = torch.tensor(
-                next(self.iterable_dataset)[self.tokens_column],
-                dtype=torch.long,
-                device=device,
-                requires_grad=False,
-            )
-            if (
-                not self.prepend_bos
-                and tokens[0] == self.model.tokenizer.bos_token_id  # type: ignore
-            ):
-                tokens = tokens[1:]
-        self.n_dataset_processed += 1
-        return tokens
+
+def validate_pretokenized_dataset_tokenizer(
+    dataset_path: str, model_tokenizer: PreTrainedTokenizerBase
+) -> None:
+    """
+    Helper to validate that the tokenizer used to pretokenize the dataset matches the model tokenizer.
+    """
+    try:
+        tokenization_cfg_path = hf_hub_download(
+            dataset_path, "sae_lens.json", repo_type="dataset"
+        )
+    except HfHubHTTPError:
+        return
+    if tokenization_cfg_path is None:
+        return
+    with open(tokenization_cfg_path, "r") as f:
+        tokenization_cfg = json.load(f)
+    tokenizer_name = tokenization_cfg["tokenizer_name"]
+    try:
+        ds_tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
+    # if we can't download the specified tokenizer to verify, just continue
+    except HTTPError:
+        return
+    if ds_tokenizer.get_vocab() != model_tokenizer.get_vocab():
+        raise ValueError(
+            f"Dataset tokenizer {tokenizer_name} does not match model tokenizer {model_tokenizer}."
+        )
