@@ -10,7 +10,7 @@ from transformer_lens.hook_points import HookedRootModule
 
 from sae_lens import __version__
 from sae_lens.config import LanguageModelSAERunnerConfig
-from sae_lens.evals import run_evals
+from sae_lens.evals import EvalConfig, run_evals
 from sae_lens.training.activations_store import ActivationsStore
 from sae_lens.training.optim import L1Scheduler, get_lr_scheduler
 from sae_lens.training.training_sae import TrainingSAE, TrainStepOutput
@@ -131,6 +131,20 @@ class SAETrainer:
         else:
             self.autocast_if_enabled = contextlib.nullcontext()
 
+        # Set up eval config
+
+        self.trainer_eval_config = EvalConfig(
+            batch_size_prompts=self.cfg.eval_batch_size_prompts,
+            n_eval_reconstruction_batches=self.cfg.n_eval_batches,
+            n_eval_sparsity_variance_batches=self.cfg.n_eval_batches,
+            compute_ce_loss=True,
+            compute_l2_norms=True,
+            compute_sparsity_metrics=True,
+            compute_variance_metrics=True,
+            compute_kl=False,
+            compute_featurewise_weight_based_metrics=False,
+        )
+
     @property
     def feature_sparsity(self) -> torch.Tensor:
         return self.act_freq_scores / self.n_frac_active_tokens
@@ -171,6 +185,12 @@ class SAETrainer:
 
             ### If n_training_tokens > sae_group.cfg.training_tokens, then we should switch to fine-tuning (if we haven't already)
             self._begin_finetuning_if_needed()
+
+        # fold the estimated norm scaling factor into the sae weights
+        if self.activation_store.estimated_norm_scaling_factor is not None:
+            self.sae.fold_activation_norm_scaling_factor(
+                self.activation_store.estimated_norm_scaling_factor
+            )
 
         # save final sae group to checkpoints folder
         self.save_checkpoint(
@@ -280,14 +300,12 @@ class SAETrainer:
         total_variance = (sae_in - sae_in.mean(0)).pow(2).sum(-1)
         explained_variance = 1 - per_token_l2_loss / total_variance
 
-        if isinstance(ghost_grad_loss, torch.Tensor):
-            ghost_grad_loss = ghost_grad_loss.item()
-        return {
+        log_dict = {
             # losses
             "losses/mse_loss": mse_loss,
             "losses/l1_loss": l1_loss
             / self.current_l1_coefficient,  # normalize by l1 coefficient
-            "losses/ghost_grad_loss": ghost_grad_loss,
+            "losses/auxiliary_reconstruction_loss": output.auxiliary_reconstruction_loss,
             "losses/overall_loss": loss,
             # variance explained
             "metrics/explained_variance": explained_variance.mean().item(),
@@ -300,6 +318,14 @@ class SAETrainer:
             "details/current_l1_coefficient": self.current_l1_coefficient,
             "details/n_training_tokens": n_training_tokens,
         }
+        # Log ghost grad if we're using them
+        if self.cfg.use_ghost_grads:
+            if isinstance(ghost_grad_loss, torch.Tensor):
+                ghost_grad_loss = ghost_grad_loss.item()
+
+            log_dict["losses/ghost_grad_loss"] = ghost_grad_loss
+
+        return log_dict
 
     @torch.no_grad()
     def _run_and_log_evals(self):
@@ -308,28 +334,35 @@ class SAETrainer:
             self.cfg.wandb_log_frequency * self.cfg.eval_every_n_wandb_logs
         ) == 0:
             self.sae.eval()
-            eval_metrics = run_evals(
+            eval_metrics, _ = run_evals(
                 sae=self.sae,
                 activation_store=self.activation_store,
                 model=self.model,
-                n_eval_batches=self.cfg.n_eval_batches,
-                eval_batch_size_prompts=self.cfg.eval_batch_size_prompts,
+                eval_config=self.trainer_eval_config,
                 model_kwargs=self.cfg.model_kwargs,
-            )
+            )  # not calculating featurwise metrics here.
 
-            W_dec_norm_dist = self.sae.W_dec.norm(dim=1).detach().float().cpu().numpy()
-            b_e_dist = self.sae.b_enc.detach().float().cpu().numpy()
+            # Remove eval metrics that are already logged during training
+            eval_metrics.pop("metrics/explained_variance", None)
+            eval_metrics.pop("metrics/explained_variance_std", None)
+            eval_metrics.pop("metrics/l0", None)
+            eval_metrics.pop("metrics/l1", None)
+            eval_metrics.pop("metrics/mse", None)
 
-            # More detail on loss.
+            # Remove metrics that are not useful for wandb logging
+            eval_metrics.pop("metrics/total_tokens_evaluated", None)
 
-            # add weight histograms
-            eval_metrics = {
-                **eval_metrics,
-                **{
-                    "weights/W_dec_norms": wandb.Histogram(W_dec_norm_dist),
-                    "weights/b_e": wandb.Histogram(b_e_dist),
-                },
-            }
+            W_dec_norm_dist = self.sae.W_dec.detach().float().norm(dim=1).cpu().numpy()
+            eval_metrics["weights/W_dec_norms"] = wandb.Histogram(W_dec_norm_dist)  # type: ignore
+
+            if self.sae.cfg.architecture == "standard":
+                b_e_dist = self.sae.b_enc.detach().float().cpu().numpy()
+                eval_metrics["weights/b_e"] = wandb.Histogram(b_e_dist)  # type: ignore
+            elif self.sae.cfg.architecture == "gated":
+                b_gate_dist = self.sae.b_gate.detach().float().cpu().numpy()
+                eval_metrics["weights/b_gate"] = wandb.Histogram(b_gate_dist)  # type: ignore
+                b_mag_dist = self.sae.b_mag.detach().float().cpu().numpy()
+                eval_metrics["weights/b_mag"] = wandb.Histogram(b_mag_dist)  # type: ignore
 
             wandb.log(
                 eval_metrics,
@@ -341,7 +374,7 @@ class SAETrainer:
     def _build_sparsity_log_dict(self) -> dict[str, Any]:
 
         log_feature_sparsity = _log_feature_sparsity(self.feature_sparsity)
-        wandb_histogram = wandb.Histogram(log_feature_sparsity.numpy())
+        wandb_histogram = wandb.Histogram(log_feature_sparsity.numpy())  # type: ignore
         return {
             "metrics/mean_log10_feature_sparsity": log_feature_sparsity.mean().item(),
             "plots/feature_density_line_chart": wandb_histogram,
