@@ -9,7 +9,7 @@ from transformer_lens import HookedTransformer
 from sae_lens.config import LanguageModelSAERunnerConfig
 from sae_lens.training.activations_store import ActivationsStore
 from sae_lens.training.sae_trainer import SAETrainer
-from sae_lens.training.training_sae import TrainingSAE
+from sae_lens.training.training_sae import JumpReLU, TrainingSAE
 from tests.unit.helpers import build_sae_cfg
 
 
@@ -19,7 +19,6 @@ from tests.unit.helpers import build_sae_cfg
         {
             "model_name": "tiny-stories-1M",
             "dataset_path": "roneneldan/TinyStories",
-            "tokenized": False,
             "hook_name": "blocks.1.hook_resid_pre",
             "hook_layer": 1,
             "d_in": 64,
@@ -27,7 +26,6 @@ from tests.unit.helpers import build_sae_cfg
         {
             "model_name": "tiny-stories-1M",
             "dataset_path": "roneneldan/TinyStories",
-            "tokenized": False,
             "hook_name": "blocks.1.hook_resid_pre",
             "hook_layer": 1,
             "d_in": 64,
@@ -37,7 +35,6 @@ from tests.unit.helpers import build_sae_cfg
         {
             "model_name": "tiny-stories-1M",
             "dataset_path": "apollo-research/roneneldan-TinyStories-tokenizer-gpt2",
-            "tokenized": False,
             "hook_name": "blocks.1.hook_resid_pre",
             "hook_layer": 1,
             "d_in": 64,
@@ -45,7 +42,6 @@ from tests.unit.helpers import build_sae_cfg
         {
             "model_name": "tiny-stories-1M",
             "dataset_path": "apollo-research/roneneldan-TinyStories-tokenizer-gpt2",
-            "tokenized": False,
             "hook_name": "blocks.1.hook_resid_pre",
             "hook_layer": 1,
             "d_in": 64,
@@ -54,7 +50,6 @@ from tests.unit.helpers import build_sae_cfg
         {
             "model_name": "tiny-stories-1M",
             "dataset_path": "roneneldan/TinyStories",
-            "tokenized": False,
             "hook_name": "blocks.1.attn.hook_z",
             "hook_layer": 1,
             "d_in": 64,
@@ -87,7 +82,7 @@ def training_sae(cfg: Any):
 @pytest.fixture
 def activation_store(model: HookedTransformer, cfg: LanguageModelSAERunnerConfig):
     return ActivationsStore.from_config(
-        model, cfg, dataset=Dataset.from_list([{"text": "hello world"}] * 2000)
+        model, cfg, override_dataset=Dataset.from_list([{"text": "hello world"}] * 2000)
     )
 
 
@@ -393,3 +388,45 @@ def test_SparseAutoencoder_set_decoder_norm_to_unit_norm(
     assert torch.allclose(
         torch.norm(sae.W_dec, dim=1), torch.ones_like(sae.W_dec[:, 0])
     )
+
+
+def test_jumprelu_forward():
+    x = torch.tensor([-1.0, 0.5, 1.5, 2.5])
+    threshold = torch.tensor(1.0)
+    expected_output = torch.tensor([0.0, 0.0, 1.5, 2.5])
+    output = JumpReLU.apply(x, threshold, 0.001)
+    assert torch.allclose(output, expected_output)  # type: ignore
+
+
+def test_jumprelu_backward():
+    x = torch.tensor([-1.0, 0.5, 1.5, 2.5], requires_grad=True)
+    threshold = torch.tensor(1.0)
+    output = JumpReLU.apply(x, threshold, 0.001)
+    output.sum().backward()  # type: ignore
+    expected_grad_x = torch.tensor([0.0, 0.0, 1.0, 1.0])
+    assert torch.allclose(x.grad, expected_grad_x)  # type: ignore
+
+
+def test_jumprelu_backward_with_threshold_grad():
+    bandwidth = 1e-3
+    threshold_value = 1.0
+    threshold = torch.tensor(threshold_value, requires_grad=True)
+
+    epsilon = 1e-8
+    x = torch.tensor(
+        [
+            threshold_value - bandwidth / 2 + epsilon,
+            threshold_value + bandwidth / 2 - epsilon,
+        ],
+        requires_grad=True,
+    )
+
+    output = JumpReLU.apply(x, threshold, bandwidth)
+    output.sum().backward()  # type: ignore
+
+    expected_grad_x = torch.tensor([0.0, 1.0])
+    expected_grad_threshold = torch.tensor(-2000.0)
+
+    atol = 1e-2
+    assert torch.allclose(x.grad, expected_grad_x, atol=atol)  # type: ignore
+    assert torch.isclose(threshold.grad, expected_grad_threshold, atol=atol)  # type: ignore
